+++ src/lib/app-provider/hooks/useConvertHexAddress.ts
@@ -1,7 +1,7 @@
 import { useCallback } from "react";
 
 import { HEX_MODULE_ADDRESS_LENGTH, HEX_WALLET_ADDRESS_LENGTH } from "lib/data";
-import type { HexAddr } from "lib/types";
+import type { BechAddr20, BechAddr32, HexAddr } from "lib/types";
 import { hexToBech32Address } from "lib/utils";
 
 import { useCurrentChain } from "./useCurrentChain";
@@ -14,12 +14,20 @@ export const useConvertHexAddress = () => {
   return {
     convertHexWalletAddress: useCallback(
       (hexAddr: HexAddr) =>
-        hexToBech32Address(bech32Prefix, hexAddr, HEX_WALLET_ADDRESS_LENGTH),
+        hexToBech32Address(
+          bech32Prefix,
+          hexAddr,
+          HEX_WALLET_ADDRESS_LENGTH
+        ) as BechAddr20,
       [bech32Prefix]
     ),
     convertHexModuleAddress: useCallback(
       (hexAddr: HexAddr) =>
-        hexToBech32Address(bech32Prefix, hexAddr, HEX_MODULE_ADDRESS_LENGTH),
+        hexToBech32Address(
+          bech32Prefix,
+          hexAddr,
+          HEX_MODULE_ADDRESS_LENGTH
+        ) as BechAddr32,
       [bech32Prefix]
     ),
   };