+++ src/lib/services/evm/index.ts
@@ -1,8 +1,18 @@
 import { useQuery } from "@tanstack/react-query";
 
-import { CELATONE_QUERY_KEYS, useLcdEndpoint } from "lib/app-provider";
+import {
+  CELATONE_QUERY_KEYS,
+  useCurrentChain,
+  useLcdEndpoint,
+} from "lib/app-provider";
+import type { HexAddr20 } from "lib/types";
+import { isHexWalletAddress } from "lib/utils";
 
-import { getEvmParams } from "./lcd";
+import {
+  getEvmCodesByAddress,
+  getEvmContractInfoSequencer,
+  getEvmParams,
+} from "./lcd";
 
 export const useEvmParams = () => {
   const lcdEndpoint = useLcdEndpoint();
@@ -15,6 +25,37 @@ export const useEvmParams = () => {
       retry: false,
       staleTime: Infinity,
       retryOnMount: false,
+    }
+  );
+};
+
+export const useEvmCodesByAddress = (address: HexAddr20) => {
+  const lcdEndpoint = useLcdEndpoint();
+
+  return useQuery(
+    [CELATONE_QUERY_KEYS.EVM_CODES_BY_ADDRESS_LCD, lcdEndpoint],
+    async () => getEvmCodesByAddress(lcdEndpoint, address),
+    {
+      retry: 1,
+      refetchOnWindowFocus: false,
+      enabled: address && isHexWalletAddress(address),
+    }
+  );
+};
+
+export const useEvmContractInfoSequencer = (address: HexAddr20) => {
+  const lcdEndpoint = useLcdEndpoint();
+  const {
+    chain: { bech32_prefix: prefix },
+  } = useCurrentChain();
+
+  return useQuery(
+    [CELATONE_QUERY_KEYS.EVM_CONTRACT_INFO_SEQUENCER, lcdEndpoint, prefix],
+    async () => getEvmContractInfoSequencer(lcdEndpoint, prefix, address),
+    {
+      retry: 1,
+      refetchOnWindowFocus: false,
+      enabled: address && isHexWalletAddress(address),
     }
   );
 };