+++ src/lib/services/types/evm/codes.ts
@@ -1,0 +1,8 @@
+import { z } from "zod";
+
+export const zEvmCodesByAddressResponseLcd = z.object({
+  code: z.string(),
+});
+export type EvmCodesByAddressResponseLcd = z.infer<
+  typeof zEvmCodesByAddressResponseLcd
+>;